+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Prevented create decoder instance until start read a message from topics
+
 ## v3.99.4
 * Fixed bug with wrong context on session closing
 * Fixed goroutine leak on closing `database/sql` driver